Fix zip stream being opened from FSP collapsed to text, which has problems with long paths, use FSP's OpenStream rather.
Fix zip entry stream not closed after use.

Switch to use of our common ZIP tool, no particular benefits, but the metadata reader on *NIX does not like it with System.IO.Compression at the moment, since it's still running on Mono and cannot find it. So it's a workaround to make it satisfied, and no real problem since it's just a good a ZIP we use everywhere.